backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
aiomysql==0.2.0
aiosqlite==0.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt<5
PyJWT==2.8.0
python-multipart==0.0.9
